worker_server/requirements.txt
pikepdf==8.11.2
requests==2.31.0
pdf2image==1.17.0
python-pptx==0.6.23
pdfplumber
pandas
openpyxl
fpdf
pdf2docx==3.1.2
pandas==2.1.4
fpdf==1.7.2
